% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_extensions.R
\name{available_extensions}
\alias{available_extensions}
\title{List of Quarto extensions in the package}
\usage{
available_extensions()
}
\value{
a string of available extensions for install
}
\description{
List of Quarto extensions in the package
}
