% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\docType{data}
\name{coral_pink}
\alias{coral_pink}
\title{Define global colours}
\format{
An object of class \code{character} of length 1.
}
\usage{
coral_pink
}
\description{
Define global colours
}
\keyword{datasets}
