% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggONS-package.R
\docType{package}
\name{ggONS-package}
\alias{ggONS}
\alias{ggONS-package}
\title{ggONS: Functions for ONS branded plots and presentations}
\description{
Provides templates and extensions for presentations with Quarto, and includes functions for styling plots and tables.
}
\author{
\strong{Maintainer}: Nicola Rennie \email{nrennie35@gmail.com}

}
\keyword{internal}
