% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\docType{data}
\name{grey_50}
\alias{grey_50}
\title{Define global colours}
\format{
An object of class \code{character} of length 1.
}
\usage{
grey_50
}
\description{
Define global colours
}
\keyword{datasets}
