% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\docType{data}
\name{lavendar_purple}
\alias{lavendar_purple}
\title{Define global colours}
\format{
An object of class \code{character} of length 1.
}
\usage{
lavendar_purple
}
\description{
Define global colours
}
\keyword{datasets}
