% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\docType{data}
\name{ocean_blue}
\alias{ocean_blue}
\title{Define global colours}
\format{
An object of class \code{character} of length 1.
}
\usage{
ocean_blue
}
\description{
Define global colours
}
\keyword{datasets}
