% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\docType{data}
\name{ONScolours}
\alias{ONScolours}
\title{List of colours.}
\format{
An object of class \code{list} of length 4.
}
\usage{
ONScolours
}
\description{
Complete list of palettes.
Use names(ONScolours) to return all palette names.
}
\keyword{datasets}
