% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\name{ONSpalettes}
\alias{ONSpalettes}
\title{ONS palettes
Generates the colour palettes}
\usage{
ONSpalettes(palette_name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{palette_name}{Name of Palette.}

\item{n}{Number of desired colours. If number of requested colours is beyond the scope
of the palette, colours are automatically interpolated.}

\item{type}{The type of palette. One of c("discrete", "continuous").}
}
\value{
A vector of colours.
}
\description{
ONS palettes
Generates the colour palettes
}
