% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\name{scale_color_ONS_c}
\alias{scale_color_ONS_c}
\title{Function to change colours in ggplot for continuous variables}
\usage{
scale_color_ONS_c(palette_name, ...)
}
\arguments{
\item{palette_name}{Name of palette.}

\item{...}{other arguments to be passed to \code{scale_colour_gradientn}}
}
\description{
Function to change colours in ggplot for continuous variables
}
