% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\name{scale_fill_ONS_d}
\alias{scale_fill_ONS_d}
\title{Function to change fill in ggplot for discrete variables}
\usage{
scale_fill_ONS_d(palette_name, ...)
}
\arguments{
\item{palette_name}{Name of palette.}

\item{...}{other arguments to be passed to \code{scale_fill_manual}}
}
\description{
Function to change fill in ggplot for discrete variables
}
