% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ONS_palette.R
\name{set_ONS_palette}
\alias{set_ONS_palette}
\title{Set ONS base R plotting palette}
\usage{
set_ONS_palette(palette)
}
\arguments{
\item{palette}{Name of palette. See \code{names(ONScolours)}. Run
\code{palette("default")} to reset.}
}
\description{
Set ONS base R plotting palette
}
