% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ONS_par.R
\name{set_ONS_par}
\alias{set_ONS_par}
\title{Set ONS base R graphical parameters}
\usage{
set_ONS_par(
  family = "Open Sans",
  adj = 0,
  mar = c(5, 3, 3, 2.5),
  bty = "n",
  ...
)
}
\arguments{
\item{family}{Font used for all text elements. Default "OPen Sans".}

\item{adj}{Alignment of text for title. Default 0.}

\item{mar}{Margins. Default \code{c(5, 3, 3, 2.5)}.}

\item{bty}{Axis lines. Default "n".}

\item{...}{Additional arguments passed to \code{par}}
}
\description{
Set ONS base R graphical parameters
}
