% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONSpalettes.R
\docType{data}
\name{spring_green}
\alias{spring_green}
\title{Define global colours}
\format{
An object of class \code{character} of length 1.
}
\usage{
spring_green
}
\description{
Define global colours
}
\keyword{datasets}
