% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ONS.R
\name{theme_ONS}
\alias{theme_ONS}
\title{Theme for \code{ggplot2} plots}
\usage{
theme_ONS(family = "Open Sans", bg_col = "transparent")
}
\arguments{
\item{family}{Name of font family to use for the text in the
plot. Default \code{"Open Sans"}.}

\item{bg_col}{Background colour of plot. Default \code{"transparent"}.}
}
\description{
Theme for \code{ggplot2} plots
}
