% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_ext.R
\name{use_quarto_ext}
\alias{use_quarto_ext}
\title{Install bundled Quarto extensions into current working directory
https://github.com/jthomasmock/octavo/blob/master/R/use_quarto_ext.R}
\usage{
use_quarto_ext(
  file_name = NULL,
  ext_name = "ONSstyle",
  use_template = TRUE,
  open = FALSE
)
}
\arguments{
\item{file_name}{Name of new qmd file}

\item{ext_name}{String indicating which extension to install}

\item{use_template}{Boolean indicating if template file should be copied}

\item{open}{Boolean indicating whether or not to open new file. Default FALSE}
}
\value{
a message if extension was successfully copied over
}
\description{
Install bundled Quarto extensions into current working directory
https://github.com/jthomasmock/octavo/blob/master/R/use_quarto_ext.R
}
